test(spli): split tests into different jobs on github actions
