+++ .github/workflows/test.yaml
@@ -21,7 +21,7 @@ jobs:
       - name: lint
         run: npm run lint
 
-  test:
+  integration tests:
     name: test
     runs-on: windows-latest
     env:
@@ -60,7 +60,7 @@ jobs:
 
       - name: test
         run: |
-          npm run test
+          npm run test:integrationTests
         env:
           BROWSERSTACK_ACCESS_KEY: ${{ secrets.BROWSERSTACK_ACCESS_KEY }}
           BROWSERSTACK_USERNAME: ${{ secrets.BROWSERSTACK_USERNAME }}
@@ -80,3 +80,123 @@ jobs:
           files: ./coverage.cobertura.xml
           fail_ci_if_error: true
           verbose: true
+
+browser tests:
+  name: test
+  runs-on: windows-latest
+  env:
+    config: 'Release'
+
+  steps:
+    - name: checkout
+      uses: actions/checkout@v3
+
+    - name: setup dotnet
+      uses: actions/setup-dotnet@v2
+      with:
+        dotnet-version: '6.0.x'
+    - name: setup node
+      uses: actions/setup-node@v3
+      with:
+        node-version: '16.x'
+    - name: setup java
+      uses: actions/setup-java@v3
+      with:
+        distribution: 'microsoft'
+        java-version: '17'
+    - name: install node deps
+      run: npm install
+    - name: node build
+      run: npm run build
+    - name: install dotnet deps
+      run: |
+        dotnet tool install -g coverlet.console
+        dotnet tool install -g dotnet-reportgenerator-globaltool
+        dotnet restore
+    - name: build
+      run: npm run dotnet:build
+    - name: start solr
+      run: ./web/solr/bin/solr start
+
+    - name: test
+      run: |
+        npm run test:browserTests
+      env:
+        BROWSERSTACK_ACCESS_KEY: ${{ secrets.BROWSERSTACK_ACCESS_KEY }}
+        BROWSERSTACK_USERNAME: ${{ secrets.BROWSERSTACK_USERNAME }}
+
+    - name: console coverage
+      if: always()
+      run: |
+        # generate a report just to show in the action history
+        reportgenerator -reports:coverage.cobertura.xml -targetdir:coverage/ -reporttypes:textSummary
+        # print out the report
+        cat coverage/Summary.txt
+
+    - name: upload cov
+      if: always()
+      uses: codecov/codecov-action@v3
+      with:
+        files: ./coverage.cobertura.xml
+        fail_ci_if_error: true
+        verbose: true
+
+function tests:
+  name: test
+  runs-on: windows-latest
+  env:
+    config: 'Release'
+
+  steps:
+    - name: checkout
+      uses: actions/checkout@v3
+
+    - name: setup dotnet
+      uses: actions/setup-dotnet@v2
+      with:
+        dotnet-version: '6.0.x'
+    - name: setup node
+      uses: actions/setup-node@v3
+      with:
+        node-version: '16.x'
+    - name: setup java
+      uses: actions/setup-java@v3
+      with:
+        distribution: 'microsoft'
+        java-version: '17'
+    - name: install node deps
+      run: npm install
+    - name: node build
+      run: npm run build
+    - name: install dotnet deps
+      run: |
+        dotnet tool install -g coverlet.console
+        dotnet tool install -g dotnet-reportgenerator-globaltool
+        dotnet restore
+    - name: build
+      run: npm run dotnet:build
+    - name: start solr
+      run: ./web/solr/bin/solr start
+
+    - name: test
+      run: |
+        npm run test:functionTests
+      env:
+        BROWSERSTACK_ACCESS_KEY: ${{ secrets.BROWSERSTACK_ACCESS_KEY }}
+        BROWSERSTACK_USERNAME: ${{ secrets.BROWSERSTACK_USERNAME }}
+
+    - name: console coverage
+      if: always()
+      run: |
+        # generate a report just to show in the action history
+        reportgenerator -reports:coverage.cobertura.xml -targetdir:coverage/ -reporttypes:textSummary
+        # print out the report
+        cat coverage/Summary.txt
+
+    - name: upload cov
+      if: always()
+      uses: codecov/codecov-action@v3
+      with:
+        files: ./coverage.cobertura.xml
+        fail_ci_if_error: true
+        verbose: true
