+++ web.Tests/FunctionTests/Helpers/Extensions.Tests.cs
@@ -3,7 +3,7 @@ using Atlas_Web.Helpers;
 using Atlas_Web.Models;
 using System;
 
-namespace web.Tests;
+namespace web.Tests.FunctionTests;
 using System.Collections.Generic;
 using System.Linq;
