+++ web.Tests/FunctionTests/Helpers/Url.Tests.cs
@@ -3,7 +3,7 @@ using Atlas_Web.Helpers;
 using Atlas_Web.Models;
 using System;
 
-namespace web.Tests;
+namespace web.Tests.FunctionTests;
 
 public class ModelHelpersTests : IClassFixture<TestDatabaseFixture>
 {
