+++ web.Tests/web.Tests.csproj
@@ -40,9 +40,6 @@
     <ProjectReference Include="..\web\web.csproj" />
   </ItemGroup>
   <ItemGroup>
-     <InternalsVisibleTo Include="web" />
-</ItemGroup>
-  <ItemGroup>
     <Content Include="..\.env*">
       <CopyToOutputDirectory>Always</CopyToOutputDirectory>
     </Content>
